Table writer 7: ID generator for distinct partitions
Add an ID generator class PartitionIdGenerator that
generates sequential integer IDs for distinct partition
values. It takes partition RowVector as input and returns
the generated IDs for the corresponding rows.

The implementation is based on VectorHasher, so rehash
could happen which will change the integer IDs for the
previous partition values that the PartitionIdGenerator was
ever applied to since the first input. In that case, a
non-empty map will be returned that the users could utilize
to map the old IDs to the new IDs after rehash. Otherwise
in the case of a stable run, an empty map will be returned.