Table writer 9: Support partition write in TableWriter
When the table is partitioned, create multiple dwrf writers
by HiveDataSink, one for each distinct partition value.
Call PartitionIdGenerator to generate sequential IDs
for distinct partition values, which are used for the index
of the dwrf writer vector.

Each input could see new partition values, so dwrf writers
are created incrementally. When PartitionIdGenerator is applied
to new input, it could cause PartitionIdGenerator to rehash,
and IDs for existing partition values change. We handle this
by recomputing the new IDs for existing partition values
and moving the dwrf writers to the new indices.

To assemble the input for different dwrf writers, we compute
the index into the original input for every partition value
and wrap the original input into a DictionaryVector.
The DictionaryVector is sent to the corresponding dwrf writer
as writer input.